+++ qp-generator-frontend/src/Views/BloomLogic.jsx
